PotegaApp/backend/requirements.txt.txt
asgiref==3.8.1
Django==4.2.17
django-cors-headers==4.6.0
djangorestframework==3.15.2
psycopg2-binary>=2.9,<3.0
sqlparse==0.5.3
tzdata==2024.2